styling changes for projects index
